変数の型を変更

+++ Assets/NetworkTest/Scripts/ConversionUtil.cs
@@ -33,25 +33,25 @@ public class ConversionUtil
 
     static public float ToFloat(byte[] values)
     {
-        int x = 0;
+        uint x = 0;
 
         if (System.BitConverter.IsLittleEndian)
         {
-            x |= (values[3] <<  0);
-            x |= (values[2] <<  8);
-            x |= (values[1] << 16);
-            x |= (values[0] << 24);
+            x |= ((uint)values[3] <<  0);
+            x |= ((uint)values[2] <<  8);
+            x |= ((uint)values[1] << 16);
+            x |= ((uint)values[0] << 24);
         }
         else
         {
-            x |= (values[0] <<  0);
-            x |= (values[1] <<  8);
-            x |= (values[2] << 16);
-            x |= (values[3] << 24);
+            x |= ((uint)values[0] <<  0);
+            x |= ((uint)values[1] <<  8);
+            x |= ((uint)values[2] << 16);
+            x |= ((uint)values[3] << 24);
         }
 
         UIntFloat uif = new UIntFloat();
-        uif.intValue = (uint)x;
+        uif.intValue = x;
         return uif.floatValue;
     }
 
@@ -64,33 +64,33 @@ public class ConversionUtil
 
     static public double ToDouble(byte[] values)
     {
-        long x = 0;
+        ulong x = 0;
 
         if (System.BitConverter.IsLittleEndian)
         {
-            x |= ((long)values[7] <<  0);
-            x |= ((long)values[6] <<  8);
-            x |= ((long)values[5] << 16);
-            x |= ((long)values[4] << 24);
-            x |= ((long)values[3] << 32);
-            x |= ((long)values[2] << 40);
-            x |= ((long)values[1] << 48);
-            x |= ((long)values[0] << 56);
+            x |= ((ulong)values[7] <<  0);
+            x |= ((ulong)values[6] <<  8);
+            x |= ((ulong)values[5] << 16);
+            x |= ((ulong)values[4] << 24);
+            x |= ((ulong)values[3] << 32);
+            x |= ((ulong)values[2] << 40);
+            x |= ((ulong)values[1] << 48);
+            x |= ((ulong)values[0] << 56);
         }
         else
         {
-            x |= ((long)values[0] <<  0);
-            x |= ((long)values[1] <<  8);
-            x |= ((long)values[2] << 16);
-            x |= ((long)values[3] << 24);
-            x |= ((long)values[4] << 32);
-            x |= ((long)values[5] << 40);
-            x |= ((long)values[6] << 48);
-            x |= ((long)values[7] << 56);
+            x |= ((ulong)values[0] <<  0);
+            x |= ((ulong)values[1] <<  8);
+            x |= ((ulong)values[2] << 16);
+            x |= ((ulong)values[3] << 24);
+            x |= ((ulong)values[4] << 32);
+            x |= ((ulong)values[5] << 40);
+            x |= ((ulong)values[6] << 48);
+            x |= ((ulong)values[7] << 56);
         }
 
         UIntFloat uif = new UIntFloat();
-        uif.longValue = (ulong)x;
+        uif.longValue = x;
         return uif.doubleValue;
     }
